games/arcade/snake-game/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
redis==5.0.1
sqlalchemy==2.0.23
aiosqlite==0.19.0
slowapi==0.1.9
python-dotenv==1.0.0
greenlet==3.0.3
httpx==0.25.2
